feat: improve ipc inheritance and code further
